requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
paramiko==3.4.0
openai==1.6.1
python-dotenv==1.0.0
requests==2.31.0
bcrypt==4.1.2

